refactor(changelog-generator): extract console functions into separate file
To declutter.
